requirements.txt
Django~=2.2.7
djangorestframework~=3.10.3
Pygments~=2.4.2
httpie~=1.0.3

